% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_columns_to_42.R
\name{add_columns_to_42}
\alias{add_columns_to_42}
\title{Add columns to 42}
\usage{
add_columns_to_42(df)
}
\arguments{
\item{df}{a data frame}
}
\description{
Internal function required by make_dam_file.R. Fills empty monitor value to reach 32 monitor columns required by DAM format (TriKinetics Drosophila Activity Monitor)
}
