% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nblobs.R
\name{filter_nblobs}
\alias{filter_nblobs}
\title{Filter to number of blobs}
\usage{
filter_nblobs(by_change)
}
\description{
Converts a "by_change" input and outputs a data frame containing total blobs/movements to be used as input for make_dam_file()
}
