% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_movements.R
\name{find_movements}
\alias{find_movements}
\title{Find movements}
\usage{
find_movements(
  files,
  n_thr = 0.996,
  n_cln = 5,
  n_grw = 1.5,
  n_blr = 3,
  n_max = 75000,
  x_left = 0,
  x_right = 2592,
  y_bot = 0,
  y_top = 1944,
  find_thr = T,
  type_thr = "absolute",
  p_sample = 0.02,
  channel = "grayscale",
  animal = "white"
)
}
\arguments{
\item{files}{a vector of file names}

\item{x_left}{a number for leftmost crop boundary}

\item{x_right}{a number for rightmost crop boundary}

\item{y_bot}{a number for bottom crop boundary}

\item{y_top}{a number for top crop boundary}
}
\description{
This function identifies all "blobs" or movements between consecutive images
}
\examples{
# out <- find_movements(files = file_names,n_max = 75000,find_thr = T,type_thr = "absolute",p_sample = 0.2,channel = "grayscale", animal = "white")

}
