% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_threshold.R
\name{find_threshold}
\alias{find_threshold}
\title{Find pixel difference cutoff for thresholding}
\usage{
find_threshold(
  files,
  p_sample = 0.05,
  n_thr = 0.999,
  n_blr = 3,
  x_left = 0,
  x_right = 2592,
  y_bot = 0,
  y_top = 1944,
  channel = "grayscale",
  animal = "white"
)
}
\description{
This function is called within find_movements()
}
\examples{
\dontrun{
 pix_diff <- find_threshold(files=files[1:(0.5*length(files))],
 p_sample=p_sample,
 n_blr=3,
 n_thr=0.998,
 x_left=x_left, x_right=x_right,
 y_bot=y_bot, y_top=y_top,
 channel=channel, animal=animal)
 }
}
