% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iLAMtools-package.R
\docType{package}
\name{iLAMtools-package}
\alias{iLAMtools}
\alias{iLAMtools-package}
\title{iLAMtools: iLAM tools}
\description{
Helpful functions to process/analyze images captured by imaging Locomotor Activity Monitors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daytonjn/iLAMtools}
  \item Report bugs at \url{https://github.com/daytonjn/iLAMtools/issues}
}

}
\author{
\strong{Maintainer}: Jacob Dayton \email{jacob.dayton@tufts.edu} (\href{https://orcid.org/0000-0002-0989-8481}{ORCID})

}
\keyword{internal}
