% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dam_file.R
\name{make_dam_file}
\alias{make_dam_file}
\title{Make DAM file}
\usage{
make_dam_file(by_frame, variable_name = "s")
}
\arguments{
\item{by_frame}{a data frame}

\item{variable_name}{a character indicating whether file should contain "s" (sum of movements) or "n" (total number of movements/blobs)}
}
\description{
Converts by_change input to DAM format (TriKinetics Drosophila Activity Monitor)
}
\examples{
# DAM_by_frame = make_dam_file(by_frame_input, variable_name = "s")

}
