% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gif.R
\name{make_gif}
\alias{make_gif}
\title{Make gif}
\usage{
make_gif(out_file_name, pi_sub_folder)
}
\arguments{
\item{out_file_name}{the name of the output gif}

\item{pi_sub_folder}{the name of the folder within ./mvmnt_\link{subfolder}}
}
\description{
This function will make a gif of all images contained with ./mvmnt_\link{subfolder}
}
\examples{
# make_gif(out_file_name, pi_sub_folder)

}
