% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_movements.R
\name{parse_movements}
\alias{parse_movements}
\title{Parse movements}
\usage{
parse_movements(file_mvmnts, start_photophase, end_photophase)
}
\arguments{
\item{file_mvmnts}{a csv containing dataframe output by find_movement()}

\item{start_photophase}{a number for the time that photophase started in standard time}

\item{end_photophase}{a number for the time that photophase ended in standard time}
}
\description{
Parse .csv containing the movements output by find_movement()
}
\examples{
\dontrun{
# by_change <- parse_movements(file_mvmnts = paste0("Onub_fhlbnd_202308",".csv"),
# start_photophase = 2,
# end_photophase = 14)
}
}
