% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_movements.R
\name{plot_movements}
\alias{plot_movements}
\title{Draw movements}
\usage{
plot_movements(
  pi_sub_folder,
  by_change,
  x_left = 0,
  x_right = 2592,
  y_bot = 0,
  y_top = 1944
)
}
\arguments{
\item{pi_sub_folder}{name of subfolder containing images}

\item{by_change}{name of dataframe containing movements (name of csv rows much correspond to image names)}

\item{x_left}{number for crop (must match crop used in find_movements())}

\item{x_right}{number for crop (must match crop used in find_movements())}

\item{y_bot}{number for crop (must match crop used in find_movements())}

\item{y_top}{number for crop (must match crop used in find_movements())}
}
\description{
This function will circle identified movements/blobs onto their corresponding image files
}
