% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iLAMtools-package.R
\docType{package}
\name{iLAMtools-package}
\alias{iLAMtools}
\alias{iLAMtools-package}
\title{iLAMtools: iLAM tools}
\description{
Helpful functions to process/analyze images captured by imaging Locomotor Activity Monitors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iLAMtools/iLAMtools}
  \item Report bugs at \url{https://github.com/iLAMtools/iLAMtools/issues}
}

}
\author{
\strong{Maintainer}: first name last name \email{email@institution.edu} (\href{https://orcid.org/anonymized for review}{ORCID})

}
\keyword{internal}
